Real Image Challenge 2016: Fixes as Suggested by Reviewer.    * Splitted the app.py to seperate the user interface from the Logic.    * Changed the authorize method(which handles the authorization) from recursion to iteration.    * Added the condition to check whether elements exists in the array or not to avoid the Index Out of Bound exception.
